// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/SystemusersModel.ts
export const Systemusersaccessmode = {
  0: 'Read_Write',
  1: 'Administrative',
  2: 'Read',
  3: 'SupportUser',
  4: 'Non_interactive',
  5: 'DelegatedAdmin'
} as const;
export type Systemusersaccessmode = keyof typeof Systemusersaccessmode;
export const Systemusersaddress1_addresstypecode = {
  1: 'DefaultValue'
} as const;
export type Systemusersaddress1_addresstypecode = keyof typeof Systemusersaddress1_addresstypecode;
export const Systemusersaddress1_shippingmethodcode = {
  1: 'DefaultValue'
} as const;
export type Systemusersaddress1_shippingmethodcode = keyof typeof Systemusersaddress1_shippingmethodcode;
export const Systemusersaddress2_addresstypecode = {
  1: 'DefaultValue'
} as const;
export type Systemusersaddress2_addresstypecode = keyof typeof Systemusersaddress2_addresstypecode;
export const Systemusersaddress2_shippingmethodcode = {
  1: 'DefaultValue'
} as const;
export type Systemusersaddress2_shippingmethodcode = keyof typeof Systemusersaddress2_shippingmethodcode;
export const Systemusersazurestate = {
  0: 'Exists',
  1: 'Softdeleted',
  2: 'Notfoundorharddeleted'
} as const;
export type Systemusersazurestate = keyof typeof Systemusersazurestate;
export const Systemuserscaltype = {
  0: 'Professional',
  1: 'Administrative',
  2: 'Basic',
  3: 'DeviceProfessional',
  4: 'DeviceBasic',
  5: 'Essential',
  6: 'DeviceEssential',
  7: 'Enterprise',
  8: 'DeviceEnterprise',
  9: 'Sales',
  10: 'Service',
  11: 'FieldService',
  12: 'ProjectService'
} as const;
export type Systemuserscaltype = keyof typeof Systemuserscaltype;
export const Systemusersdeletedstate = {
  0: 'Notdeleted',
  1: 'Softdeleted'
} as const;
export type Systemusersdeletedstate = keyof typeof Systemusersdeletedstate;
export const Systemusersemailrouteraccessapproval = {
  0: 'Empty',
  1: 'Approved',
  2: 'PendingApproval',
  3: 'Rejected'
} as const;
export type Systemusersemailrouteraccessapproval = keyof typeof Systemusersemailrouteraccessapproval;
export const Systemusersincomingemaildeliverymethod = {
  0: 'None',
  1: 'MicrosoftDynamics365forOutlook',
  2: 'Server_SideSynchronizationorEmailRouter',
  3: 'ForwardMailbox'
} as const;
export type Systemusersincomingemaildeliverymethod = keyof typeof Systemusersincomingemaildeliverymethod;
export const Systemusersinvitestatuscode = {
  0: 'InvitationNotSent',
  1: 'Invited',
  2: 'InvitationNearExpired',
  3: 'InvitationExpired',
  4: 'InvitationAccepted',
  5: 'InvitationRejected',
  6: 'InvitationRevoked'
} as const;
export type Systemusersinvitestatuscode = keyof typeof Systemusersinvitestatuscode;
export const Systemusersoutgoingemaildeliverymethod = {
  0: 'None',
  1: 'MicrosoftDynamics365forOutlook',
  2: 'Server_SideSynchronizationorEmailRouter'
} as const;
export type Systemusersoutgoingemaildeliverymethod = keyof typeof Systemusersoutgoingemaildeliverymethod;
export const Systemuserspreferredaddresscode = {
  1: 'MailingAddress',
  2: 'OtherAddress'
} as const;
export type Systemuserspreferredaddresscode = keyof typeof Systemuserspreferredaddresscode;
export const Systemuserspreferredemailcode = {
  1: 'DefaultValue'
} as const;
export type Systemuserspreferredemailcode = keyof typeof Systemuserspreferredemailcode;
export const Systemuserspreferredphonecode = {
  1: 'MainPhone',
  2: 'OtherPhone',
  3: 'HomePhone',
  4: 'MobilePhone'
} as const;
export type Systemuserspreferredphonecode = keyof typeof Systemuserspreferredphonecode;
export const Systemuserssystemmanagedusertype = {
  0: 'EntraUser',
  1: 'C2User',
  2: 'ImpersonableStubUser',
  3: 'AgenticUser'
} as const;
export type Systemuserssystemmanagedusertype = keyof typeof Systemuserssystemmanagedusertype;

export interface SystemusersBase {
  accessmode: Systemusersaccessmode;
  address1_addressid?: string;
  address1_addresstypecode?: Systemusersaddress1_addresstypecode;
  address1_city?: string;
  address1_country?: string;
  address1_county?: string;
  address1_fax?: string;
  address1_latitude?: number;
  address1_line1?: string;
  address1_line2?: string;
  address1_line3?: string;
  address1_longitude?: number;
  address1_name?: string;
  address1_postalcode?: string;
  address1_postofficebox?: string;
  address1_shippingmethodcode?: Systemusersaddress1_shippingmethodcode;
  address1_stateorprovince?: string;
  address1_telephone1?: string;
  address1_telephone2?: string;
  address1_telephone3?: string;
  address1_upszone?: string;
  address1_utcoffset?: number;
  address2_addressid?: string;
  address2_addresstypecode?: Systemusersaddress2_addresstypecode;
  address2_city?: string;
  address2_country?: string;
  address2_county?: string;
  address2_fax?: string;
  address2_latitude?: number;
  address2_line1?: string;
  address2_line2?: string;
  address2_line3?: string;
  address2_longitude?: number;
  address2_name?: string;
  address2_postalcode?: string;
  address2_postofficebox?: string;
  address2_shippingmethodcode?: Systemusersaddress2_shippingmethodcode;
  address2_stateorprovince?: string;
  address2_telephone1?: string;
  address2_telephone2?: string;
  address2_telephone3?: string;
  address2_upszone?: string;
  address2_utcoffset?: number;
  applicationid?: string;
  azurestate: Systemusersazurestate;
  "BusinessUnitId@odata.bind": string;
  "CalendarId@odata.bind"?: string;
  caltype: Systemuserscaltype;
  displayinserviceviews?: boolean;
  domainname: string;
  emailrouteraccessapproval: Systemusersemailrouteraccessapproval;
  employeeid?: string;
  entityimage?: string;
  firstname: string;
  governmentid?: string;
  homephone?: string;
  importsequencenumber?: number;
  incomingemaildeliverymethod: Systemusersincomingemaildeliverymethod;
  internalemailaddress: string;
  invitestatuscode: Systemusersinvitestatuscode;
  isallowedbyipfirewall?: boolean;
  isdisabled?: boolean;
  isintegrationuser: boolean;
  islicensed: boolean;
  issyncwithdirectory: boolean;
  jobtitle?: string;
  lastname: string;
  middlename?: string;
  mobilealertemail?: string;
  "MobileOfflineProfileId@odata.bind"?: string;
  mobilephone?: string;
  nickname?: string;
  outgoingemaildeliverymethod: Systemusersoutgoingemaildeliverymethod;
  overriddencreatedon?: string;
  "ParentSystemUserId@odata.bind"?: string;
  passporthi?: number;
  passportlo?: number;
  personalemailaddress?: string;
  photourl?: string;
  "PositionId@odata.bind"?: string;
  preferredaddresscode?: Systemuserspreferredaddresscode;
  preferredemailcode?: Systemuserspreferredemailcode;
  preferredphonecode?: Systemuserspreferredphonecode;
  processid?: string;
  "QueueId@odata.bind"?: string;
  salutation?: string;
  setupuser: boolean;
  sharepointemailaddress?: string;
  skills?: string;
  stageid?: string;
  systemmanagedusertype: Systemuserssystemmanagedusertype;
  systemuserid: string;
  "TerritoryId@odata.bind"?: string;
  timezoneruleversionnumber?: number;
  title?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  traversedpath?: string;
  userlicensetype: number;
  utcconversiontimezonecode?: number;
  windowsliveid?: string;
  wsfmv_externalagilinkuserid?: string;
  yammeremailaddress?: string;
  yammeruserid?: string;
  yomifirstname?: string;
  yomilastname?: string;
  yomimiddlename?: string;
}

export interface Systemusers extends SystemusersBase {
  accessmodename?: string;
  activedirectoryguid?: string;
  address1_addresstypecodename?: string;
  address1_composite?: string;
  address1_shippingmethodcodename?: string;
  address2_addresstypecodename?: string;
  address2_composite?: string;
  address2_shippingmethodcodename?: string;
  applicationiduri?: string;
  azureactivedirectoryobjectid?: string;
  azuredeletedon?: string;
  azurestatename?: string;
  businessunitidname: string;
  caltypename?: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  defaultfilterspopulated: boolean;
  defaultmailboxname?: string;
  defaultodbfoldername: string;
  deletedstate: Systemusersdeletedstate;
  deletedstatename?: string;
  disabledreason?: string;
  displayinserviceviewsname?: string;
  emailrouteraccessapprovalname?: string;
  entityimage_timestamp?: number;
  entityimage_url?: string;
  entityimageid?: string;
  exchangerate?: number;
  fullname?: string;
  identityid: number;
  incomingemaildeliverymethodname?: string;
  invitestatuscodename?: string;
  isactivedirectoryuser: boolean;
  isallowedbyipfirewallname?: string;
  isdisabledname?: string;
  isemailaddressapprovedbyo365admin: boolean;
  isintegrationusername?: string;
  islicensedname?: string;
  latestupdatetime?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  organizationid: string;
  organizationidname: string;
  outgoingemaildeliverymethodname?: string;
  parentsystemuseridname?: string;
  parentsystemuseridyominame?: string;
  positionidname?: string;
  preferredaddresscodename?: string;
  preferredemailcodename?: string;
  preferredphonecodename?: string;
  queueidname?: string;
  setupusername?: string;
  systemmanagedusertypename?: string;
  territoryidname: string;
  transactioncurrencyidname?: string;
  userpuid?: string;
  versionnumber?: number;
  yomifullname?: string;
  businessunitid?: object;
  _businessunitid_value?: string;
  calendarid?: object;
  _calendarid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  defaultmailbox?: object;
  _defaultmailbox_value?: string;
  mobileofflineprofileid?: object;
  _mobileofflineprofileid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  parentsystemuserid?: object;
  _parentsystemuserid_value?: string;
  positionid?: object;
  _positionid_value?: string;
  queueid?: object;
  _queueid_value?: string;
  territoryid?: object;
  _territoryid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}